# DR-3454 Snapshot/Dataset Retrieve Query Refactors