This PR implements the query proposal from this doc:
https://docs.google.com/document/d/1vmp81uoqS_qtswT0Jo7y76y8GRZwAdGqzRd4XiVyPP8/edit#heading=h.pjsggzrj5awt

This is to reduce load on the database.  This PR aims to reduce 1:N queries (e.g. loops that call sql queries).

While I haven't been able to really observe the load benefit in the wild, I can say that for individual requests (e.g. get dataset and get snapshot) benefits:

Before fix:
- Get dataset: ~400ms
- Get Snapshot: ~650ms

After fix:
- Get dataset: ~300ms
- Get Snapshot: ~300ms

So I expect to get some benefit just on perf in addition to taking load off of the DB which will help in general.

The areas that I focused on were:
- Combining table and column queries into one for datasets and snapshots and
- Combining the query that build mappings for snapshots into one

Oh, and I also converted some the other bigger queries in the various daos into multiline strings